Restructured to enable usage of profiling func outside of class
