+++ petk/explore.py
@@ -1,8 +1,5 @@
 from collections import OrderedDict
 
-from shapely.validation import explain_validity
-
-import importlib
 import warnings
 
 import geopandas as gpd
@@ -11,12 +8,7 @@ import pandas as pd
 import matplotlib
 
 import petk.constants as constants
-import petk.utils as utils
-
-has_geopy = not importlib.util.find_spec("geopy") is None
-if has_geopy:
-    from geopandas.tools import reverse_geocode
-    from shapely.geometry import MultiPoint
+import petk.tools as tools
 
 
 class DataReport:
@@ -47,13 +39,13 @@ class DataReport:
 
         additions.append(
             pd.Series([
-            ('columns', '{0}'.format(utils.get_type(self.df[col]).lower())) for col in self.df.columns
+            ('columns', '{0}'.format(tools.get_type(self.df[col]).lower())) for col in self.df.columns
             ]).value_counts()
         )
 
         if isinstance(self.df, gpd.GeoDataFrame):
             has_z = self.df.has_z.value_counts()
-            centroid_loc = ', '.join(reverse_geocode(MultiPoint(self.df.centroid).centroid)['address'][0].split(', ')[1:]) if has_geopy else None
+            centroid_loc = tools.get_point_location(MultiPoint(self.df.centroid))
 
             additions.append(
                 pd.Series({
@@ -92,7 +84,7 @@ class DataReport:
 
         for c in columns:
             if c not in self.description.columns:
-                self.description = pd.concat([self.description, self.get_description(c)], axis=1, sort=False)
+                self.description = pd.concat([self.description, tools.get_description(self.df[c], name=c)], axis=1, sort=False)
 
         return self.description[columns]
 
@@ -115,22 +107,22 @@ class DataReport:
                 warnings.warn('Column {0} not found in data'.format(column), stacklevel=2)
                 continue
 
-            dtype = utils.get_type(self.df[column])
+            dtype = tools.get_type(self.df[column])
 
             if dtype in [constants.TYPE_GEO]:
                 for func, kwargs in rule.items():
                     if isinstance(kwargs, bool) and kwargs:
                         kwargs = {}
 
-                    result = getattr(self, func)(**kwargs)
+                    result = getattr(tools, func)(self.df, **kwargs)
 
-                    if result:
+                    if result is not None:
                         validation[func] = result
             else:
                 pass
 
         if validation:
-            validation = pd.concat(validation.gdfvalues(), keys=validation.keys(), sort=False).reset_index()
+            validation = pd.concat(validation.values(), keys=validation.keys(), sort=False).reset_index()
             validation.columns = ['function', 'index', 'notes']
 
             # Sort the MultiIndex by the record index then the issue found
@@ -140,153 +132,3 @@ class DataReport:
                 validation = validation.join(self.df, how='inner')
 
         return validation
-
-    def get_description(self, column):
-        '''
-        Profile a single data column
-
-        Parameters:
-        columns (str): Identify the column to profile
-
-        Returns:
-        (OrderedDict): Profile result
-        '''
-
-        series = self.df[column]
-
-        count = series.count() # ONLY non-NaN observations
-        dtype = utils.get_type(series)
-
-        description = {
-            'type': dtype,
-            'column': series.name,
-            'memory_usage': series.memory_usage(),
-            'count': count,
-            'p_missing': (series.size - count) / series.size,
-            'n_missing': series.size - count,
-        }
-
-        if not dtype in [constants.TYPE_UNSUPPORTED, constants.TYPE_CONST, constants.TYPE_UNIQUE, constants.TYPE_GEO]:
-            n_distinct = series.nunique()
-
-            description.update({
-                'distinct_count': n_distinct,
-                'is_unique': n_distinct == series.size,
-                'p_unique': n_distinct * 1.0 / series.size
-            })
-
-            if dtype == constants.TYPE_BOOL:
-                description.update({
-                    'mean': series.mean()
-                })
-            elif dtype in [constants.TYPE_DATE, constants.TYPE_NUM]:
-                n_inf = series.loc[(~np.isfinite(series)) & series.notnull()].size
-
-                description.update({
-                    'p_infinite': n_inf / series.size,
-                    'n_infinite': n_inf,
-                    'min': series.min(),
-                    'max': series.max()
-                })
-
-                for perc in [0.05, 0.25, 0.5, 0.75, 0.95]:
-                    description['{:.0%}'.format(perc)] = series.quantile(perc)
-
-                if dtype == constants.TYPE_NUM:
-                    n_zeros = series.size - np.count_nonzero(series)
-
-                    description.update({
-                        'mean': series.mean(),
-                        'std': series.std(),
-                        'variance': series.var(),
-                        'iqr': series.quantile(0.75) - series.quantile(0.25),
-                        'kurtosis': series.kurt(),
-                        'skewness': series.skew(),
-                        'sum': series.sum(),
-                        'mad': series.mad(),
-                        'cv': series.std() / series.mean(),
-                        'n_zeros': n_zeros,
-                        'p_zeros': n_zeros / series.size
-                    })
-
-        # OrderedDict used to fixed the DataFrame column orders
-        return pd.DataFrame(pd.Series(description, name=column))
-
-    def get_invalids(self):
-        '''
-        Find the invalid geometries within the data
-
-        Returns:
-        (DataFrame): Invalid geometries and the reason
-        '''
-
-        invalids = self.df[~self.df.is_valid]
-
-        if not invalids.empty:
-            return invalids['geometry'].apply(lambda x: explain_validity(x) if not x is None else 'Null geometry')
-
-    def get_outsiders(self, xmin, xmax, ymin, ymax):
-        '''
-        Find the geometries outside of a certain bounding box
-
-        Parameters:
-        xmin (numeric): Min X of the bounding box
-        xmax (numeric): Max X of the bounding box
-        ymin (numeric): Min Y of the bounding box
-        ymax (numeric): Max Y of the bounding box
-
-        Returns:
-        (DataFrame): Out of bound geometries
-        '''
-
-        assert xmin < xmax and ymin < ymax, 'Invalid bounding box'
-
-        # TODO: There has to be a better way to select outside of a bounding box...
-        outsiders = self.df.loc[~self.df.index.isin(self.df.cx[xmin:xmax, ymin:ymax].index)]
-
-        if not outsiders.empty:
-            return outsiders['geometry'].apply(lambda x: 'Geometry outside of bbox({0}, {1}, {2}, {3})'.format(xmin, xmax, ymin, ymax))
-
-    def get_slivers(self, projection=None, area_thresh=constants.SLIVER_AREA, length_thresh=constants.SLIVER_LINE):
-        '''
-        Find the slivers within the geometries
-
-        Parameters:
-        projection    (numeric): EPSG number for the geometry of the data
-        area_thresh   (numeric): Threshold area for a Polygon to be considered a sliver
-        length_thresh (numeric): Threshold length for a Line to be considered a sliver
-
-        Returns:
-        (DataFrame): Sliver geometries
-        '''
-
-        assert projection, 'Projection required'
-
-        def is_sliver(x, area_thresh, length_thresh):
-            if 'polygon' in x.geom_type.lower():
-                return x.area < area_thresh
-            elif 'linestring' in x.geom_type.lower():
-                return x.length < length_thresh
-            else:   # Points
-                return False
-
-        pieces = self.df.explode().to_crs({'init': 'epsg:{0}'.format(projection), 'units': 'm'})
-
-        slivers = pieces['geometry'].apply(is_sliver, args=(area_thresh, length_thresh))
-        slivers = slivers[slivers].groupby(level=0).count()
-
-        if not slivers.empty:
-            return slivers.apply(lambda x: '{0} slivers found within geometry'.format(x))
-
-    # def get_frequent_values(self, top):
-    #     return series.value_counts(dropna=False).head(top)
-    #
-    # def get_distribution(self, top):
-    #     assert dtype in [constants.STR, constants.DATE, constants.NUM], 'Distribution not available'
-    #
-    #     if dtype == constants.STR:
-    #         # Histogram of most frequen top values
-    #         pass
-    #     else:
-    #         # Distribution Plot
-    #         pass
